backend/requirements.txt
Django>=4.0,<5.0
django-ninja
mysqlclient
django-cors-headers
pydantic[email]
